share post component working
