+++ src/shared/components/post/PostBody.tsx
@@ -1,6 +1,7 @@
 import { Box, Typography } from "@mui/material";
 import useTheme from "features/theme/useTheme";
 import { BasePostBodyType } from "shared/types";
+import { getSecureUrl } from "shared/utils/cloudinaryUtil";
 
 function PostBody({ content, media }: BasePostBodyType) {
   const {
@@ -9,19 +10,24 @@ function PostBody({ content, media }: BasePostBodyType) {
   return (
     <>
       <Box component="section" sx={{ mt: "15px" }}>
-        <Typography sx={{fontSize: "16px", color}}>
-          {content}
-        </Typography>
+        <Typography sx={{ fontSize: "16px", color }}>{content}</Typography>
       </Box>
       <Box
         component="div"
         sx={{ display: "flex", mt: "20px", justifyContent: "center" }}
       >
-        <Box
-          component="img"
-          sx={{ borderRadius: "10px", width: '100%', height: '100%', display: 'block' }}
-          src={media}
-        ></Box>
+        {media && (
+          <Box
+            component="img"
+            sx={{
+              borderRadius: "10px",
+              width: "100%",
+              height: "100%",
+              display: "block",
+            }}
+            src={getSecureUrl(media)}
+          />
+        )}
       </Box>
     </>
   );
